test: Adjust test for Stan 2.35
